chore: move getCollections outside
