+++ src/processors/base-processor/processor.ts
@@ -1,9 +1,6 @@
 import {Anonymizer} from '../../anonymizers/types';
-import {Config} from '../../config/types';
-import {Collection} from '../types/collection';
 
 export interface Processor {
-	getCollections(config: Config): Collection[];
 	processColumn(
 		tableName: string,
 		columnName: string,
